requirements.txt
fastapi
uvicorn
python-magic
python-multipart
jinja2
